Bump version to 1.4.3 and improve log handling
Updated the application version from 1.4.2 to 1.4.3 in the Dockerfile and main.py. Enhanced the log retrieval functionality in the `Owner` class by improving error handling and streamlining the process of sending log files. Introduced a new variable for the log file path to enhance readability, and refined the zipping and sending of logs. Improved command argument handling for robustness and enhanced the logic for retrieving and sending the last X lines of the log file.